add header

+++ c_srcs/client.c
@@ -1,27 +1,25 @@
-#include <signal.h>
-#include <unistd.h>
-#include "../libft/srcs/libft.h"
-
-//void signal_handler()
-//{
-//
-//}
+#include "../includes/minitalk.h"
 
 int main(int argc, char **argv)
 {
 	int		s_pid;
-	int		message;
 	size_t	i;
+	size_t	j;
 
-	s_pid = ft_atoi(argv[1]);
-	i = 0;
-	while (i < ft_strlen(argv[2]))
+	j = 0;
+	if (argc <= 2)
 	{
-		while (j < 8)
+		s_pid = ft_atoi(argv[1]);
+		i = 0;
+		while (i < ft_strlen(argv[2]))
 		{
-			kill(s_pid, argv[2][j]);
-			j++;
+			while (j < 8)
+			{
+				kill(s_pid, argv[2][j]);
+				j++;
+			}
+			i++;
 		}
-		i++;
 	}
+	return (0);
 }
